feat: ArtImg, Art, ArtPost 테스트 추가
